% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{before_after}
\alias{before_after}
\title{before after}
\usage{
before_after(x, FUN, ...)
}
\arguments{
\item{x}{A vector}

\item{f}{function}
}
\description{
comparison tool for evaluating original values and returned function values
}
\examples{
before_after()
}
\keyword{after;}
\keyword{before;}
\keyword{comparison}
