% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{clean_date}
\alias{clean_date}
\title{clean date}
\usage{
clean_date(x, format = "monthday")
}
\arguments{
\item{format}{valid input includes 'monthday'}
}
\description{
Clean date data
}
\examples{
clean_date()
}
\keyword{date}
