% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{clean_string}
\alias{clean_string}
\title{clean string}
\usage{
clean_string(x, set_case = NULL, set_recode = NULL, FUN = NULL, ...)
}
\arguments{
\item{set_case}{valid input can be 'title','lower','upper'}

\item{set_recode}{input should be a named vector. ex: "remove char" = "replace char"}

\item{FUN}{an additionanl function call}
}
\description{
Clean string data
}
\examples{
clean_string()
}
\keyword{string}
