% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\docType{data}
\name{get_regex}
\alias{get_regex}
\title{common regex}
\format{An object of class \code{list} of length 7.}
\usage{
get_regex
}
\description{
Returns common regex text
}
\examples{
get_regex()
}
\keyword{regex}
