% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{is_valid}
\alias{is_valid}
\title{Is Valid Values}
\usage{
is_valid(x)
}
\arguments{
\item{x}{}
}
\description{
This function selects only non-null, non-na values, non-blank, and non text 'NA' values.
}
\examples{
is_valid()
}
\keyword{valid}
