% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{remove_ws}
\alias{remove_ws}
\title{Removes whitespace.}
\usage{
remove_ws(x)
}
\arguments{
\item{x}{A vector}
}
\description{
Removes any leading or trailing whitespace. Removes more than one consecutive whitespace.
}
\examples{
remove_ws()
}
\keyword{whitespace}
