% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{return_values}
\alias{return_values}
\title{return values}
\usage{
return_values(index, invalid = NULL)
}
\arguments{
\item{index}{numeric list of list elements to return}

\item{invalid}{}
}
\description{
returns values of x that are within the provided numeric index
}
\examples{
return_values()
}
\keyword{return}
