% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{text_case}
\alias{text_case}
\title{Updates text case.}
\usage{
text_case(x, set_case = NULL)
}
\arguments{
\item{x}{A vector}

\item{set_case}{arguments can be "lower","upper" or "title"}
}
\description{
Changes the case of input text.
}
\examples{
text_case()
}
\keyword{case}
