% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{totitle}
\alias{totitle}
\title{Title case text.}
\usage{
totitle(x, width = 100)
}
\arguments{
\item{x}{A vector}

\item{width}{provide max length of string for whitespace padding purposes}
}
\description{
Capitalizes the first letter of each word. If all caps, will lower the string first.
}
\examples{
totitle()
}
\keyword{valid}
