% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{valid_date}
\alias{valid_date}
\title{valid date}
\usage{
valid_date(x, format = "all", allow = NULL, na_to_false = FALSE)
}
\arguments{
\item{format}{valid input includes 'all','monthday'}
}
\description{
Identify date data that does not meet Niche's expectations
}
\examples{
valid_date()
}
\keyword{date}
