% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{valid_numeric}
\alias{valid_numeric}
\title{valid numeric}
\usage{
valid_numeric(x, minn = 0, maxx = .Machine$integer.max, decimals = 2,
  na_to_false = FALSE)
}
\description{
Identify numeric data that does not meet Niche's expectations
}
\examples{
valid_numeric()
}
\keyword{numeric}
