% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{valid_string}
\alias{valid_string}
\title{valid string}
\usage{
valid_string(x, distinct_vals = NULL, set_regex = NULL,
  na_to_false = FALSE)
}
\description{
Identify string data that does not meet expectations
}
\examples{
valid_string()
}
\keyword{distinct}
\keyword{string}
