% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{valid_url}
\alias{valid_url}
\title{valid url}
\usage{
valid_url(x, na_to_false = FALSE)
}
\arguments{
\item{x}{A url}

\item{na_to_false}{}
}
\description{
Identify url data that does not meet expectations
}
\examples{
valid_url()
}
\keyword{valid}
