% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{valid_url_status}
\alias{valid_url_status}
\title{Finds bad urls.}
\usage{
valid_url_status(x)
}
\arguments{
\item{x}{A url}
}
\description{
Returns false for bad urls
}
\examples{
valid_url_status()
}
\keyword{valid}
